requirements.txt
requests
pandas
python-dotenv
termcolor
google-generativeai